Rename messages components
